commit Dominik Mierzejewski's patch
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@9191 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ Gui/bitmap.h
@@ -1,11 +1,10 @@
-
 #ifndef __MYSAMPLE
 #define __MYSAMPLE
 
 typedef struct _txSample
 {
- unsigned int  Width;
- unsigned int  Height;
+ unsigned long Width;
+ unsigned long Height;
  unsigned int  BPP;
  unsigned long ImageSize;
  char *        Image;
